Separated out all entitis into a pkg, separated datasources and their interfaces and New() funcs and aggregated them into the same Repository interface
