+++ datasource/rest/netbox.go
@@ -1,0 +1,3 @@
+package rest
+
+
